Some cleanings
- build.sh correctly cleans all built objects before building
- fixes: resolved warnings messages while compiling on 64bit
- removed automaticaly generated files by configure

+++ modules/libape-spidermonkey.c
@@ -2101,7 +2101,7 @@ APE_JS_NATIVE(ape_sm_set_timeout)
 		params->argv[i] = argv[i+2];
 	}
 	
-	timer = add_timeout(ms, ape_sm_timer_wrapper, params, (unsigned int)cx, g_ape);
+	timer = add_timeout(ms, ape_sm_timer_wrapper, params, cx, g_ape);
 	timer->protect = 0;
 	params->timer = timer;
 	
@@ -2145,7 +2145,7 @@ APE_JS_NATIVE(ape_sm_set_interval)
 		params->argv[i] = argv[i+2];
 	}
 	
-	timer = add_periodical(ms, 0, ape_sm_timer_wrapper, params, (unsigned int)cx, g_ape);
+	timer = add_periodical(ms, 0, ape_sm_timer_wrapper, params, cx, g_ape);
 	timer->protect = 0;
 	params->timer = timer;
 	
@@ -3000,7 +3000,7 @@ static void free_module(acetables *g_ape) // Called when module is unloaded
 
 		/* Clear all timers */
 		struct _ticks_callback *tc;
-		while ((tc = get_first_timer((unsigned int)asc->cx, g_ape)) != NULL) {
+		while ((tc = get_first_timer(asc->cx, g_ape)) != NULL) {
 			struct _ape_sm_timer *params = (struct _ape_sm_timer *)tc->params;
 			int last = 1;
 			params->cleared = 1;
